requirements.txt
discord.py==2.3.2
python-dotenv==1.0.1
APScheduler==3.10.4
matplotlib==3.9.0
Pillow==10.3.0
pytz==2024.1
numpy==1.26.4
